WIP generic scan, value types
